finished lab

+++ routes/noteRoutes.js
@@ -9,12 +9,10 @@ import { authMiddleware } from '../utils/auth.js'
 router.use(authMiddleware);
  
 // GET /api/notes - Get all notes for the logged-in user
-// THIS IS THE ROUTE THAT CURRENTLY HAS THE FLAW
 router.get('/', async (req, res) => {
-  // This currently finds all notes in the database.
-  // It should only find notes owned by the logged in user.
   try {
-    const notes = await Note.find({});
+    const notes = await Note.find(
+        {user: { $eq: req.user._id}});
     res.json(notes);
   } catch (err) {
     res.status(500).json(err);
@@ -24,10 +22,12 @@ router.get('/', async (req, res) => {
 // POST /api/notes - Create a new note
 router.post('/', async (req, res) => {
   try {
-    const note = await Note.create({
+    const note = new Note({
       ...req.body,
-      // The user ID needs to be added here
+      user: req.user._id
     });
+    console.log("inside the note:" + note)
+    await note.save();
     res.status(201).json(note);
   } catch (err) {
     res.status(400).json(err);
@@ -37,8 +37,10 @@ router.post('/', async (req, res) => {
 // PUT /api/notes/:id - Update a note
 router.put('/:id', async (req, res) => {
   try {
-    // This needs an authorization check
-    const note = await Note.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const note = await Note.findByIdAndUpdate(req.params.id, req.body, {new: true});
+    if (req.user._id != note.user) {
+        return res.status(403).json({ message: 'User forbidden from updating this note' });
+    }
     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
@@ -51,11 +53,14 @@ router.put('/:id', async (req, res) => {
 // DELETE /api/notes/:id - Delete a note
 router.delete('/:id', async (req, res) => {
   try {
-    // This needs an authorization check
-    const note = await Note.findByIdAndDelete(req.params.id);
+    const note = await Note.findById(req.params.id);
+    if (req.user._id != note.user) {
+        return res.status(403).json({ message: 'User forbidden from deleting this note' });
+    }
     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
+    const deleteNote = await Note.findByIdAndDelete(req.params.id);
     res.json({ message: 'Note deleted!' });
   } catch (err) {
     res.status(500).json(err);
